Rename Calendar class and move to a separate package
